MB-6059: Fix quoting so wrapper can handle spaces in path
Scripts generated with wrapper.in wouldn't run if the path to the
script contained spaces (as it always will on Mac OS.)

Also, the value '@prefix@' was being used in several places even
though it gets stored into a shell variable $prefix. I fixed this
so the Mac installer script install-couchbase.sh only has to fix
it in one place.

Change-Id: Iea936cc0e6226c72b627d55b13ef213ef9434bf1
Reviewed-on: http://review.couchbase.org/19052
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Jens Alfke <jens@couchbase.com>