MB-6090: source_node fallback if user used localhost/127.0.0.1
If we can't filter the source nodes because the user supplied
'localhost' or '127.0.0.1', just fallback to using all the source
nodes.

This has no effect for today's couchbase, which spreads a bucket
across all nodes.  One day, if couchbase supports assigning buckets to
less than all nodes in a cluster, this will need revisiting.

Change-Id: Ibd8911e82ced57fcb367f168b05830e420dc5529
Reviewed-on: http://review.couchbase.org/19244
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Bin Cui <bin.cui@gmail.com>