MB-6470: handle restore to hostname instead of IP address.
The filter_bucket_nodes() helper function is used to favor restoring
design-docs to the same server that the user specified.  Like...

  ./cbrestore /backup-dir http://10.3.4.5:8091

Design-docs would be restored to 10.3.4.5, even though it might not be
the first node listed in the cluster.

However, it doesn't handle a case when the user specifies a hostname
(like 'couchbase-02'), because a hostname won't match with the IP
addresses which ns_server usually returns.

In that case, this fix just picks the first node in the cluster to
restore design-docs instead of just error'ing.

Change-Id: I35dc2dc90832148cdbd5e015b28a5a74622b79e7
Reviewed-on: http://review.couchbase.org/20295
Reviewed-by: Pavel Paulau <pavel.paulau@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>