MB-7341: Check vbucket state before querying kv table
In DGM case, sqlite db file is fragmented and it is time consuming
to even one query to retrieve kv items.
Based on the fact that all items in one kv table share the same
vbucket state value, we can decide to skip the whole kv table
or query it solely based on vbucket id and vbucket state.

Change-Id: If6ec8adf238591fd9f1187e186da78da353a4947
Reviewed-on: http://review.couchbase.org/23205
Reviewed-by: Chisheng Hong <chisheng@couchbase.com>
Tested-by: Bin Cui <bin.cui@gmail.com>