MB-16878: Re-arrange the cluster-init calls to set the REST credentials last
Once the REST credentials are set the cluster is fully initialized
so we should do this last. Otherwise we can run into a problem where
a REST call that happens later in the process fails and causes the
user to no longer be able to use the cluster-init command again.

Change-Id: I287ec010c797cb8b23369ceb854bdb99b38916cd
Reviewed-on: http://review.couchbase.org/58077
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>