MB-16152: Add the initial version of the cluster manager
One of the problems contributing to code bloat and complexity is
the absence of a proper REST interface for Couchbase. This class
intends to clean up this complexity by abstracting away all REST
calls to Couchbase.

The initial version adds a proper way to query the N1QL service and
takes into account the fact that the cluster might not be homogenous.

Change-Id: I3d44ffa602ab35411a8057107578011147f1595d
Reviewed-on: http://review.couchbase.org/58765
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>