Don't catch Exception when loading indexes from the docloader
It is very bad practice to catch Execption in python because this
will mask programming errors. For now we will catch the only
exception we know about and can add others later when we see them
occur. It's better to get a stack trace than to ignore a compile
error anyways.

Change-Id: I5f1bfd9af09b3ca49865c01ca66290e9b4aec4a9
Reviewed-on: http://review.couchbase.org/58924
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>