MB-16654: Never use the rest username/password as the sasl credentials
Buckets have their own username and password that are separate from
the rest credentials. If we cannot get the bucket username/password
from the source_bucket variable then something has gone wrong long
before we get to the point where we try to use it.

Note that this doesn't fix any issue. It just removes some logic
that is incorrect.

Change-Id: I242a2973fb859651d6bc9a64332815cf34bfb232
Reviewed-on: http://review.couchbase.org/59356
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>