Suppress the mode flag from the CE cbbackup help message
The mode flag is not used for the CE version, but we still list it. In
order to make sure we don't break backwards compatibility and that
the mode option is still added to the list internally let's just
supporess the help message for the mode flag in CE mode.

Change-Id: Ia431f441c5146790e65b8935e2f29120b5d6d8cf
Reviewed-on: http://review.couchbase.org/60658
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>