Refactor json loader in order to improve performance
The current json loader would extract the json zip file twice in
order to load data and this was problematic because on Windows it
can take up to 3.5 minutes to unzip the file.

The new code does not extract the contents of the zip file and
instead reads directly from the zip files.

Change-Id: I764d0538dedea69adc60e5b587a1ab2f7722e359
Reviewed-on: http://review.couchbase.org/60988
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>