Remove the cbclusterstats command line tool
This is a tool that hasn't been updated in over 3 years and was
originally developed to get stats from production clusters for
debugging purposes. The code is old and this project never really
made it out into a Couchbase release. We will remove the code for
now and if we do need it in the future we can always revert this
change.

Change-Id: Ibf4837597ff9b0a1db1ddec25385a28fda31cc5f
Reviewed-on: http://review.couchbase.org/63144
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>