MB-20776: Deprecate failing over multiple servers at the same time
For some reason servers['failover'] is a dictionary while only
a single element that looks something like the struct below.

{'127.0.0.1:8091', True}

As a result we can always look at the first key to get the failover
names. Since we are re-writing couchbase-cli in the next release
we can leave this ugliness for now.

Change-Id: I9d9e079ab714787099f91b459a0fc0c745e9e05c
Reviewed-on: http://review.couchbase.org/68378
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>