MB-21422: Allow index storage setting to be set without the index service
This change allows us to set the index storage mode if we are not
adding the index service during cluster-init. Note that if we do
add the index service that we want to default to forestdb based
indexes if the --index-storage-setting parameter is not set. If they
are not setting up the index service then we won't set the storage
mode unless the parameter is explicitly specified.

Change-Id: Id57fe5a81998926d232616455640caf4f158804f
Reviewed-on: http://review.couchbase.org/69090
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>