Allow command line arguments to have optional values
When we add a command line argument for a password we want to be
able to either have the password specified on the command line or
if there is no password given to be able to prompt the user for
the password using non-echoed stdin. In order to do this with the
getopt library we will replace some of the parsing function to allow
us to add a new parameter type to getopt. This parameter type is
specified with as "<param>?".

Note that this implementation is a hack and that it will be replaced
in the next release when we start using the argparse library.

Change-Id: I4a74f24a6f98945675d00bd8437cabdf0269bace
Reviewed-on: http://review.couchbase.org/70144
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>