MB-21778: Disallow restoring a seqno backup to an lww backup
If we are doing conflict resolution with our data then we cannot
restore a backup that was backed up with a different conflict
resolution type then the one we are restoring to. The excpetion here
is if we are forcing updates using either the conflict-resolve or
set_xwm options because these options skip conflict resolution.

Change-Id: Iac90eaf893ba579a6d225833b04f2db76ccef2dc
Reviewed-on: http://review.couchbase.org/70414
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>