MB-21735: Ensure we get the executable path on windows
On windows the os.path.dirname function doesn't expand the path to
a full path. I'm assuming that this is because on linux we run the
command with ./ and this indicates the current directory, but on
windows we don't do this. As a result we should explicitly expand
the absolute path.

Change-Id: Idebdfe8fc706b393cf8d4f0a6ec0b7c1df68bc6b
Reviewed-on: http://review.couchbase.org/70445
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>