MB-22239: Add the ability to confirm passwords through non-echoed stdin
Some password that are entered through non-echoed stdin should be
confirmed to ensure that the user did not make a typo when entering
the password.

We have added an optional confirm_text variable to the CBNonEchoedAction.
When this parameter is set we will prompt the user with the text in
passed to this variable to confirm their password.

Change-Id: I26f5762c31aa8065bb669cb64b0dd6110c22a8dd
Reviewed-on: http://review.couchbase.org/72208
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>