Support plasma indexes in couchbase-cli
We are changing the default index storage mode to plasma and will
deprecate the forestdb backend. If the forestdb backend is already
choosen then we should continue to use it if the user specifies that
they want to use the default storage type. If they are not using
forestdb based indexes then we need to use the plasma backend.

Change-Id: I8b5595eea71c3b5c9efab9a91aa31c71fa5a8bfa
Reviewed-on: http://review.couchbase.org/75926
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>