MB-24022: Don't hang if get no response
In the case where a nodes underlying disk is not available we still
want to be able to run cbtransfer to collect data from all the nodes
where the disk is still accessible.

Currently cbtransfer hangs as it gets no response from the node where
the underlying disks are not available.  When using TAP it does not
hang.

This patch introduces a timeout of a minimum of 30 seconds.  If a
response is not received before the timeout expires then we mark
the collection of a dcp batch as done, allowing the thread to complete
and the cbtransfer tool to not hang.

Change-Id: Ice98eb8c095aaec09bf44c1fdaae5d0a28213c63
Reviewed-on: http://review.couchbase.org/77113
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Daniel Owen <owend@couchbase.com>