MB-23914: The server version for each node should be persisted during backup
We can use the server version to determine various restrictions that
we need to be aware of when backuping up different versions of
Couchbase. In particular the addition of xattrs in Couchbase 5.0
means that we need to make sure that the next backup we do is a
full backup. If a backup doesn't have a version number we set the
version number to 0.0.0 and assume that the backup was taken against
a pre-5.0 Couchbase cluster.

Change-Id: I1dd9e90cafd51dacdcd734fee38c9356360784be
Reviewed-on: http://review.couchbase.org/77415
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>