MB-25910: Properly remap the vbucketmap during cbrecovery
I made a mistake about how the vbucket map works and this caused
some issues with cbrecovery locating the right server to send
messages to. The vbucket map only contains a list for where the
active and replica vbuckets are an the first entry in the list is
always the active vbucket. We should update that entry with the
index of the server to send operations to when the index corresponds
to an entry in the serverList.

Change-Id: Idb24034839fa80006093e8ce7e6f9b262591d484
Reviewed-on: http://review.couchbase.org/83177
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>