MB-27689: Backport "Track the sqlite version during restore"
We have different versions of the way we store data in sqlite and
we need to make sure we properly track the version number thoughout
the restore in order to make sure we query the database with the
right query.

Without tracking the version we will default to the most basic query
which defaults certain fields like the datatype to 0. Doing this
will cause issues for xattr documents.

Change-Id: I7676ff46f4b139555ed2d48630b2817c18e89572
Reviewed-on: http://review.couchbase.org/88216
Reviewed-by: Patrick Varley <patrick@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Mike Wiederhold <mike@couchbase.com>