Unbreak --vm babysitter/couchdb option to server-eshell.
Until recently, the babysitter cookie (that is used to connect both to
couchdb and babysitter vms) consisted only of alphabetical
characters. Now it's a hex-encoded random byte string. Which means
that it may start with a digit. Atoms in erlang need to be enclosed in
single quotes if they start with a digit (among other things). The
get_babysitter_node() would consider these single quotes part of the
cookie. To prevent this, the corresponding /diag/eval is changed to
return the cookie as a json string. The _post_form_encoded will
automatically decode the result into a python string (based on the
Content-Type: application/json returned by the server).

Change-Id: I52946b747acc221a63bde73a1aa8d96c59f75bb3
Reviewed-on: http://review.couchbase.org/97248
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>