MB-32179 Add support for LDAP and groups in cli
Added ability to manage user groups to the user-manage command.
Changed the previously named settings-ldap to settings-sasl, and
the settings-ldap command now manages the global LDAP settings for
the cluster.

The user-manage command now has options to manage user groups as well
as to add and remove users from groups.

User roles is no longer required to create a user, instead roles OR
group is required.

Documentation has been updated accordingly so has been the tests.

Change-Id: Ieb3f99c900b8ecdeadc83cdf1e8a2d2294e8ee39
Reviewed-on: http://review.couchbase.org/102537
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>