MB-25987 Fix SSL options on cbbackup and cbrestore
Currently the SSL options in the commands that use pump_transfer where
not using SSL even when the option was enabled.

The changes introduced should allow users to use the --ssl option
to enable secure connection, if the host given uses https ssl will
be used. Together with ssl the option --no-ssl-verify is available to
skip server certificate verification. You can provide the ca to use
for the server verification using the --cacert option.

This interface is consistent with the couchbase-cli.

Change-Id: Ibb699411cd76629b972acee8e4e68e12738f3620
Reviewed-on: http://review.couchbase.org/102724
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>