MB-32363 CSV source correct provides the doc
Before this commit CSV source (cbtransfer importing a csv) would
import the meta data about the doc (cas, rev, etc) inside the
value it self. When it should have just been importing the value
itself. It also fixes the fact that not all of the meta data was
being import correctly.

Change-Id: Ie1e8503fe7f4fb63c8d3b262b89b8b99e029e7c6
Reviewed-on: http://review.couchbase.org/102853
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>
Tested-by: Patrick Varley <patrick@couchbase.com>