MB-322285 Allow entry of secondary credentials
When transferring from CB server to CB server at the moment
cbtransfer will only allow the user to give one set of
credentials so if the source and destination servers don't have the
same credentials the transfer will fail. To fix this the following
flags have been added:
-U, --username-dest - destination cluster username
-P, --password-dest - destination cluster password

If provided to cbtranfser it will use this credentials for the
destination CB server, otherwise it will use the -u and -p.

Change-Id: Ib5f09682e3093cd34355762649f87c12d0c5314a
Reviewed-on: http://review.couchbase.org/102983
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>