MB-32529 Add enable dp mode command to CLI
Added the new command enable-developer-preview which will enable
developer preview mode. Before enabling it will let the user know
that is an irreversible process and ask for confirmation.
The output of the command looks as follows:
>> couchbase-cli enable-developer-preview -c localhost:8091 /
    -u Administrator -p asdasd --enable

>>Developer preview can not be disabled once it is enabled. If you
enter developer preview mode you will not be able to upgrade.
DO NOT USE IN PRODUCTION.
Are you sure [y/n]:

Also added a --list option that will let the user know if the cluster
is in developer preview mode already.

Change-Id: Id3b81dd3be2ec70143256cd7ac2617176275e482
Reviewed-on: http://review.couchbase.org/104018
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>