MB-33133 pump_mc support for transactions
The changes to the sink means that for cbrestore any ATRs or txn
client records will not be transferred as well as striping any txn
metadata. The same applies to cbtransfer.

Added a --force-txn flag to cbrestore and cbtransfer when provided
all data will be sent to couchbase sink including any txn data. This
option is hidden and should not be used as the are no guarantees that
the txn data is complete or that it will work as desired on restore.

Also by default cbworkloadgen has it set to true as this will avoid
the checking of xattrs for txn data as cbworkloadgen doe snot give
data with txns.

Change-Id: Id05eb3d02d41557df7c7cd17cea21b08cfbdce4d
Reviewed-on: http://review.couchbase.org/105454
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>