MB-33261 Add type hints and use bytes
Every time a document is sent via the network we were checking if
it was a string or bytes. This has a cost, to avoid this the data is
now always sent as bytes and only flipped to strings when needed.

To help ensure that all functions use the correct data types type
hints have been added where possible.

Change-Id: Iec0e5f1c80d39813e9fe1350b875685378d88ecc
Reviewed-on: http://review.couchbase.org/105862
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>