MB-34645: Update helo to return list of enabled features
The KV HELO protocol requires the client send a set of features
and in return KV sends the set of features it enabled, the client
should compare the request and response sets and handle any
difference.

This commit updates cb_bin_client and pump so that missing features
are an error.

Change-Id: Ib094264bdb88e2f76885117ac1f4c96ff764a944
Reviewed-on: http://review.couchbase.org/110829
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>