MB-31926 Consistent behaviour for graceful failover
This commit ensures that a graceful failover for a node with data
and one without ends in the same result which is that node being
brought down but not ejected of the cluster. If the node has data
then it will be moved to the other nodes before being brought down.

Change-Id: I598e6bc3041773ff30ed5fc0076aac4a9eefddd9
Reviewed-on: http://review.couchbase.org/111353
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Toby Wilds <toby.wilds@couchbase.com>