MB-34149: Handle master password correctly.
With the new erlang distribution mechanism implemented in cluster manager,
the master password command in the CLI stopped working. This change
passes in the necessary arguments to the erlang VM that will be started
by the CLI to set the master password.

Earlier the CLI was scraping the stdout of the sub-process to determine
the success/failure of the command. Now, the CLI will use the discrete
exit codes returned by babysitter's helper API to determine success/failure
of the command.

This change is dependent on the following change:
http://review.couchbase.org/#/c/111937/

Change-Id: I7b531bf26c385f0632ec3793a5ef22f958c26aec
Reviewed-on: http://review.couchbase.org/111938
Reviewed-by: Patrick Varley <patrick@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>