DOC-5647 Update couchbase-cli nav
This change seeks to update the nav for Couchbase Server to include
the latest `couchbase-cli` commands.

== Mod #1
The `couchbase-cli` commands that appear in the Couchbase Server
left-hand navigation are determined by what's listed in
docs/modules/cli/pages/_partials/cbcli/nav.adoc.

Commands added to nav:
* `couchbase-cli collection-manage`
* `couchbase-cli enable-developer-preview`
* `couchbase-cli ip-family`
* `couchbase-cli master-password`
* `couchbase-cli node-to-node-encryption`
* `couchbase-cli setting-alternate-address`
* `couchbase-cli setting-master-password`
* `couchbase-cli setting-on-demand`
* `couchbase-cli setting-query`
* `couchbase-cli setting-rebalance`
* `couchbase-cli setting-saslauthd`
* `couchbase-cli user-change-password`

== Mod #2
The commands pages themselves are located in
/docs/modules/cli/pages/cbcli. Some of these are considered internal
tools and are not meant to be listed in the nav (though, they are
still published on the site and can be reached via URL if you know
where to look). To avoid confusion about having a different number of
pages vs. number of listings in the nav file, I've added listings in
the nav for these items, but commented them out with an explanation.

Commands added to nav but commented out:
* `couchbase-cli help`
* `couchbase-cli server-eshell`

== Mod #3
At this time, it is understood that there was only one command that
was removed in 6.5: `couchbase-cli admin-role-manage`. The deprecation
notice indicated that the functionality of this command was inherited
by `couchbase-cli user-manage`, so a page alias was added to that file
to redirect the old page to the new one.

Change-Id: I01ec4519b515f103fc1ac8001264fc428908dac0
Reviewed-on: http://review.couchbase.org/113214
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>