DOC-5647 Migrate tools documentation to couchbase-cli repo
This change migrates some old and existing content from the
docs-server repo to the couchbase-cli repo.

== PREFACE

In http://review.couchbase.org/#/c/111151/, the following pages were
added to the couchbase-cli repo as the single source of truth:
* cbbackup.adoc
* cbbackupwrapper.adoc
* cbrecovery.adoc
* cbrestore.adoc
* cbrestorewrapper.adoc
* cbtransfer.adoc
* cbworkloadgen.adoc

As such, the following files have been removed from docs-server:
* cbworkloadgen-tool.adoc
* cbtransfer-tool.adoc
* cbrestore-tool.adoc
* cbrecovery.adoc
* cbbackup-tool.adoc
* cbbackup-wrapper.adoc

== SUMMARY OF CHANGES

As part of this change, I copied the files that were deleted from
docs-server and put them in /docs/modules/cli/pages/cbtools/old. This
is because I'm unsure if there is still applicable content that needs
to be added to the man pages from these files. (Once we've confirmed
all of the applicable content has been harvested, we can delete them
again from this rep0.)

In addition, some other non-manpage content was moved to
/docs/modules/cli/pages/cbtools/guides since it is largely related and
needs to be updated in conjunction with any of the these tools.

One last thing to note is that I added page aliases to the new files
from http://review.couchbase.org/#/c/111151/ so that the old pages
redirect to the new pages on the docs site.

See https://github.com/couchbase/docs-server/pull/704 for the
corresponding changes in docs-server.

Change-Id: I09d5ea003e08f3292685c49dcdd3b06681d46da9
Reviewed-on: http://review.couchbase.org/113223
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>