MB-39481 Allow enabling all events and users
The way to enable all events and users is by passing the empty string
('') to the --disable-events --disabled-users options. This will
clear the respective disable list in ns_server. This did not work
before because the request parameters 'disabled' and 'disabledUsers'
because we did a falsy check on the value and only send it if it was
truthy which the empty string is not. Now we check if the value is
None and send it otherwise.

Change-Id: I3f85b3ab966a644d812d274a7888e77a303f9251
Reviewed-on: http://review.couchbase.org/c/couchbase-cli/+/128704
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>