MB-39220 Failover separates hard and force options
Before --force would also cause a hard failover with unsafeness
parameter set. This is not the expected behaviour it should just
trigger a normal hard failover.

Now a hard failover is control by the '--hard' flag and if there
is need for the unsafeness parameter to be set that is controlled
by the '--force' flag.

Change-Id: I48cce31e711888ba4392ccf687ed5bfea4cae8a4
Reviewed-on: http://review.couchbase.org/c/couchbase-cli/+/129178
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>