MB-40339 Reduce code smelliness
The tool sonar qube has deemed this codebase 'smelly' based on a lot
of breaches of Python coding standards. To reduce the 'smelliness' of
the codebase all variables and function names now use snake case, the
number of redundant else and elif has been reduced as well as unused
variables and invalid indentation.

Change-Id: Ia65341c62c4c544311c73bb9efeae61fd3d7cfae
Reviewed-on: http://review.couchbase.org/c/couchbase-cli/+/132217
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>