MB-42056 List collections will list everything
The --list-collections now takes an optional comma separated list of
scopes for which to print the collections. If non is provided all
collections will be printed. If one or more of the scopes does not
exist a warning will be printed.

Example output:

Scope test:
    - b
    - a
Scope _default:
    - _default

Example JSON output:

{
    "test": [
        "b",
        "a"
    ],
    "_default": [
        "_default"
    ]
}

Change-Id: I407dfa4762403c16e8ef28fce8b6e31bb10260e8
Reviewed-on: http://review.couchbase.org/c/couchbase-cli/+/138185
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>