MB-39036 Allow ephemeral eviction policies in bucket-edit
This has the effect of passing validation onto the REST API meaning the
correct error message will be returned when attempting to update the
eviction policy for an ephemeral bucket.

$ couchbase-cli bucket-edit -c ${ADDRESS} -u ${USERNAME} -p ${PASSWORD} --bucket ${BUCKET} --bucket-eviction-policy nruEviction
ERROR: evictionPolicy - Eviction policy cannot be updated for ephemeral buckets

NOTE: We purposefully don't document that 'noEviction' and 'nruEviction'
are accepted options since they are technically not valid. We simply
want to provide the user with a more informative error message.

Change-Id: Ida7bf366c4a160ac6b448f1514bb1dd791e3bf0d
Reviewed-on: http://review.couchbase.org/c/couchbase-cli/+/141956
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>