MB-47483 Correctly handle multi-node REST requests
When 'clusterEncryptionLevel' is set to 'strict', the non-tls ports will
be unavailable unless using 'localhost'. This will now be detected and
we'll connect using the exposed tls ports.

NOTE: We will display a warning indicating that we're having to
communicate using the tls ports and that the user may be required to
supply '--cacert' or '--no-verify-ssl'.

Change-Id: Ic92658df9b95622791469031441d4b13541f1077
Reviewed-on: http://review.couchbase.org/c/couchbase-cli/+/157868
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>