MB-50703 Correctly check for compression in sinks
If a document is compressed the second bit of the datatype will be set,
but in some places we were checking whether the whole value was
greater than two. This resulted in csv files with compressed data in
them.

NOTE: Also use constants for checking the datatype rather than literals
in all the sinks/sources.

Change-Id: I04310c91edcdce96ada8a090957871f773055f13
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/169894
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>