MB-50736 Improve server Otp names matching with different hostnames
Before this change we had several issues with matching supplied nodes'
hostnames with nodes on the cluster for several couchbase-cli
subcommands (e.g. failover, readd):

1. The nodes could not be supplied with schemes (e.g. http://[node]),
which was an unreasonable restriction given that we are not using them
to match the supplied nodes with nodes on the cluster.

Solution: trim schemes from supplied nodes since those are not
necessary for matching with nodes on the cluster.

2. SSL ports could not be used to match the supplied nodes with nodes on
the cluster. This is, again, an unreasonable restriction given that both
ports are valid in this case.

Solution: Extend matching to also consider the HTTPS port of each node
on the cluster.

3. 'localhost' clusters were not fully supported by the relevant commands
due to external hostnames of nodes and their internal Otp hostnames not
always matching.

Solution: Use both the external hostname and the internal Otp hostname
of a node on the cluster to try to match it to the supplied node.

Change-Id: I658753575d1f91c5c53ae5fef926825356621c7f
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/170361
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>