MB-45371 Don't fail when (un)deploying (un)deployed func
The previous commit for this MB changed the endpoint we use to
(un)deploy eventing functions. This new endpoint will error if you try
to deploy a deployed function or undeploy an undeployed function, which
the old endpoint did not. To retain backwards compatability we catch
this error and print we are in this state instead.

Change-Id: Icb8c0bd09b70fca80ce50a8529155d7c14f4412c
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/172096
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>