MB-51835 Fix 'master-password' and 'server-eshell' crashing server
The issue has been introduced as a result of a backport (MB-49557),
in later versions we use call 'set_cookie/1', which is not implemented
in Erlang OTP-22 that is used in 'cheshire-cat' versions. The solution
to this is to call 'set_cookie/2' instead.

The incorrect call to 'set_cookie/1' is done for two CLI sub-commands:
'master-password' and 'server-eshell', and this fix addresses both.

Change-Id: Iad2317e83a313db68d53fbd0da1f5d93afb42870
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/173808
Well-Formed: Restriction Checker
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>