MB-55677 Fix the eshell error with -d switch
These two functions:
  - get_babysitter_cookie
  - get_ns_server_cookie

are not using a tuple as param like other rest API calls, but they are
using payload instead. When the -d switch is mentioned in the eshell,
this will raise an error because the payload is sent to
_url_encode_params function where it expects param to be a tuple.
So here we are trying to avoid that function call by checking if
the parameter is actually a tuple. If not just a simple print works
fine.

Change-Id: I4973fe195902e1e12e2960f9fdf1d724b6614597
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/187269
Well-Formed: Restriction Checker
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>