MB-57219 Move from urlib3.contrib.pyopenssl to ssl
The contrib library is now deprecated so we need to move to something
else. The deprecation warning links to
https://github.com/urllib3/urllib3/issues/2680 which says we should move
to the Python standard library package 'ssl'. Unfortunately for us it
does not support PKCS#12 files and does not support loading certs/keys
from memory.

To get round this we now are forced to convert into a common format and
then write out to a temporary file. The file is deleted as soon as ssl
has read it. We take care to encrypt the private key if it was already
encrypted so this should not be a security vulnerability.

This approach was chosen as we did not want to vendor the library
ourselves and ultimately we do need to bump the dependencies which is
not possible without moving away from the contrib package as it requires
an older version of cryptography than the latest.

Change-Id: Ie02d07e0c2c2f8508c1d4744f1eb71fb54d40884
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/193583
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>