MB-58219 Create venv in fixed place
"dir" isn't used to create the venv in the Jenkins script. This causes
the attempts to activate the venv later in the script to fail, since
the correct path isn't used.

This is remedied by installing the venv in "${WORKSPACE}/venv", and
updating the references to the venv to there.

Change-Id: Ib5ca15b4926b9254b3ab05cad6f20de3929bd50b
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/195554
Tested-by: Safian Ali <safian.ali@couchbase.com>
Reviewed-by: Matt Hall <matt.hall@couchbase.com>