MB-59198 Add `--user-dn-advanced` to ldap-setting
This new parameter takes a JSON file as an argument with a list of
objects. The objects all have a match field (a regex) and either a query
or a transform field.

This field, and the existing `--user-dn` fields, are mutually exclusive
so reflect that when creating the arguments with argparse.

Change-Id: Ib8443279cd2d2348047d093cacaceea74dd2bd58
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/199242
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Safian Ali <safian.ali@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>