MB-59613: Wait until nodes != [] for bucket-create --wait
The nodes list for a bucket is not immediately populated. This means
that if things are running slowly, then we may see the bucket-create
command return early when --wait has been supplied. This is because it
will see an empty list and the logic assumes that if every node in that
list is healthy then the bucket is ready.

Since a bucket should always be on at least one node, we should
wait until the nodes list is non-empty before returning, when --wait
was supplied.

Change-Id: I124f1edaa8322130bc2ed25756d9fef2bb7d00c7
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/200737
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Hall <matt.hall@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>