MB-59852 Revert "MB-58973 Remove backup/transfer tools"
This reverts commit 9e4ab80780b0475614a8c5e4bf50db710c46e040.

QE require cbtransfer for some of their tests and would like us to add
it back for now. The hope is that once the tests are moved over (or we
have a high enough pass rate) we can remove it again. For that reason
I've reverted the entire change which includes cbbackup &co, in the hope
that we will revert the revert later.

Change-Id: I21298a4f8de4e5fe517edd3e019bf9891e64661f
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/201706
Reviewed-by: Nour Elmenyawi <nour.elmenyawi@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>