MB-59948 Allow notifications to be changed on >=7.2.3 EE
We previously added a check to ensure we only allow changing these
settings on EE clusters (or CE ones before 7.2.3). The check was correct
but we used a variable `self.enterprise` that was not being populated.
To fix we just tell the `rest_initialiser` decorator to populate it.

Change-Id: I7fc820268053c9c457f82ba533e03272c866f9a0
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/202071
Reviewed-by: Safian Ali <safian.ali@couchbase.com>
Reviewed-by: Nour Elmenyawi <nour.elmenyawi@couchbase.com>
Reviewed-by: Maksimiljans Januska <maks.januska@couchbase.com>
Tested-by: Matt Hall <matt.hall@couchbase.com>
Well-Formed: Restriction Checker