MB-58294 Version string are compared as strings
This commit implements a function called compare_versions that takes in
two strings and an equality sign, compares them and returns a true or
false based on the outcome of the equality.

For example:
v1 = 7.1.2
v2 = 10.2.3
equality = greater

compare_versions(v1,v2,equality) would return false. As v1 is not
greater than v2.

Change-Id: I1e8eb1437881ca14631e8b6a1d3541fa4fc6db2e
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/202076
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Hall <matt.hall@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>
Reviewed-by: Safian Ali <safian.ali@couchbase.com>