MB-59935 Replace use-bucket-ttl with no-expiry
ns_server has decided to keep the collection manifest APIs backwards
compatible. Because of this, the "--use-bucket-ttl" flag has been
removed, since that behaviour hasn't changed (can be done with
"--max-ttl 0"), although the docs have been changed to make this a bit
clearer.

In its place, "--no-expiry" has been added, which is equivalent to
"--max-ttl -1". Once MB-59772 is merged, this will disable TTL for all
documents in the collection.

Change-Id: I2e2f1b49d8f9b7596dabbbb51bbf3556982b7a52
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/202082
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Hall <matt.hall@couchbase.com>