MB-68462:[BP] Fix eshell after migrating to erlang 26
We set cookie in code passed via -eval and I guess there is a race
between executing -eval and -remsh. If remsh is executed first,
then cookie will not be set by the time of connect.

This fix explicitly sets the order for those operarations:
we set the cookie first, and then start the remote shell.

Change-Id: Ie9cd9ee6b5561d2e07ee25b57f3db3787bcf42eb
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/234795
Reviewed-by: Matt Hall <matt.hall@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Peter Searby <peter.searby@couchbase.com>