MB-68545[BP]: Fix eshell after migrating to erlang 26
Note: this BP depends on Erlang/OTP 26, so should ideally be merged at
the same time that as the tlm change to upgrade to OTP 26.

Original commit message below:
-----------------------------

We set cookie in code passed via -eval and I guess there is a race
between executing -eval and -remsh. If remsh is executed first,
then cookie will not be set by the time of connect.

This fix explicitly sets the order for those operarations:
we set the cookie first, and then start the remote shell.

Change-Id: Ie9cd9ee6b5561d2e07ee25b57f3db3787bcf42eb
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/233928
Reviewed-by: Matt Hall <matt.hall@couchbase.com>
Tested-by: Safian Ali <safian.ali@couchbase.com>
Well-Formed: Restriction Checker