MB-61049: Fix eshell after migrating to erlang 26
We set cookie in code passed via -eval and I guess there is a race
between executing -eval and -remsh. If remsh is executed first,
then cookie will not be set by the time of connect.

This fix explicitly sets the order for those operarations:
we set the cookie first, and then start the remote shell.

Change-Id: Ie9cd9ee6b5561d2e07ee25b57f3db3787bcf42eb
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/206836
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Safian Ali <safian.ali@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>